refactor(protocol): promote worker run contracts
