+++ packages/protocol/AGENTS.md
@@ -31,7 +31,7 @@ src/
 ├── tool-selection/       # ToolSelection schema for choosing tool categories and overrides
 ├── trace/                # TraceContext schema shared by observability helpers
 ├── worker-bootstrap/     # Worker bootstrap payload contracts
-└── subagent/             # ChildSession / WorkerRun / ConsultationRequest / BackgroundTask + Subagent.Events.*
+└── worker-run/           # WorkerRun.Info / Status + WorkerRun.Events.*
 ```
 
 ## KEY PATTERNS
@@ -42,7 +42,7 @@ src/
 - **Sink interface**: Plain TS interface (NOT Zod) — the callback contract for streaming results. Uses `Tool.Call`, `Tool.Result`, `Run.Snapshot`.
 - **BaseEvent correlation**: All events extend `BaseEvent` with `traceId`, `runId?`, `taskId?`, `sessionId?`, `time`.
 - **Policy timings**: 14 policy timing points — `inbound.receive`, `run.start`, `turn.start`, `context.prepare`, `resources.prepare`, `model.request`, `model.response`, `invoke.prepare`, `invoke.result`, `turn.finish`, `completion.prepare`, `writeback.commit`, `run.finish`, `error`. `Policy.PolicyDecision` verdict is one of `allow | deny | pending`; legacy permission evaluators still use `EvaluationResult.action` (`continue | abort`).
-- **Subagent lifecycle**: `Subagent.Events.*` covers worker sessions (`WorkerSessionSpawned/Resumed/Cancelled`), worker runs (`WorkerRunStarted/Completed/Failed`), consultations (`WorkerConsultationRequested/Completed`), and background tasks (`BackgroundTaskLaunched/Completed/Failed/Cancelled`).
+- **WorkerRun lifecycle**: `WorkerRun.Events.*` covers delegated run start, completion, failure, and cancellation.
 - **Storage sub-adapters**: `Storage.WorkItemSubAdapter` in `storage/index.ts` — pure interface contract with no runtime logic. Implementation lives in `@openomni/session`.
 - **WorkItem namespace**: `work-item/index.ts` is the public facade. `work-item/schemas.ts` defines `WorkItem.Info` (universal work state schema with derived status, routing/session fields, completion reports, retry/outcome fields, blockers, evidence, read-back checks, verification gates), `WorkItem.ReadBackCheck` (URL fetch / API query / citation match observations), and related schemas. `work-item/events.ts` defines `WorkItem.Events.*` (Created, Updated, StatusChanged, Completed, Failed, Removed via `BusEvent.define()`), `work-item/status.ts` defines `deriveStatus()` (pure status derivation from timestamps/blockers), and `work-item/hash.ts` defines `generateHash()` (base36 12-char collision-safe IDs).
 - **Execution/IPC contracts**: `execution/`, `ipc/`, and `worker-bootstrap/` describe worker requests, responses, and bootstrap payloads only. Runtime worker lifecycle lives in `@openomni/coordinator`.
@@ -74,7 +74,7 @@ Keep these as protocol contracts only. Runtime policy and storage implementation
 - Adding a new tool state? Add to `Tool.State` discriminated union in `tool/index.ts`.
 - Adding a new run type? Add to the `Run` namespace in `run/index.ts`.
 - Adding a new policy timing? Update `Policy.Timing` in `policy/index.ts` and coordinate with `packages/agent/src/core/policy/engine.ts`.
-- Adding a new subagent event? Extend `Subagent.Events` in `subagent/index.ts` with a `BusEvent.define()` call.
+- Adding a new worker-run event? Extend `WorkerRun.Events` in `worker-run/index.ts` with a `BusEvent.define()` call.
 - Adding a new storage sub-adapter interface? Add it to `storage/index.ts` as a named interface under the `Storage` namespace.
 - Adding a work-item field? Update `WorkItem.Info` in `work-item/schemas.ts`. If it affects status derivation, update `deriveStatus()` in `work-item/status.ts`.
 - Adding a work-item event? Extend `WorkItem.Events` in `work-item/events.ts` with a `BusEvent.define()` call.
