+++ packages/protocol/src/index.ts
@@ -15,7 +15,7 @@ export * from "./ingress/index.js";
 export * from "./policy/index.js";
 export * from "./agent/index.js";
 export * from "./artifact/index.js";
-export * from "./subagent/index.js";
+export * from "./worker-run/index.js";
 export * from "./execution/index.js";
 export * from "./ipc/index.js";
 export * from "./worker-bootstrap/index.js";
