+++ packages/protocol/src/policy/point-contract.ts
@@ -21,12 +21,10 @@ export namespace PolicyPointContractModule {
     "connection.llm.post",
     "tool.native.pre",
     "tool.mcp.pre",
-    "delegation.subagent.pre",
-    "delegation.background.pre",
+    "delegation.worker.pre",
     "tool.native.post",
     "tool.mcp.post",
-    "delegation.subagent.post",
-    "delegation.background.post",
+    "delegation.worker.post",
     "run.turn.post",
     "run.completion.pre",
     "session.writeback.pre",
