+++ packages/protocol/src/policy/point-registry.ts
@@ -119,19 +119,11 @@ export namespace PolicyPointRegistryModule {
       ],
       ...preBoundary,
     ),
-    "delegation.subagent.pre": contract(
-      "delegation.subagent.pre",
+    "delegation.worker.pre": contract(
+      "delegation.worker.pre",
       "pre",
       ["worker"],
-      ["sessionId", "runId", "subagentId", "subagentProfile"],
-      ["delegation.set_constraints", "delegation.require_approval", "run.abort", "audit.annotate"],
-      ...preBoundary,
-    ),
-    "delegation.background.pre": contract(
-      "delegation.background.pre",
-      "pre",
-      ["worker"],
-      ["sessionId", "runId", "backgroundTaskId"],
+      ["sessionId", "runId", "workerRunId", "workerProfile"],
       ["delegation.set_constraints", "delegation.require_approval", "run.abort", "audit.annotate"],
       ...preBoundary,
     ),
@@ -151,19 +143,11 @@ export namespace PolicyPointRegistryModule {
       ["audit.annotate", "run.abort", "tool.rewrite_output"],
       ...postBoundary,
     ),
-    "delegation.subagent.post": contract(
-      "delegation.subagent.post",
+    "delegation.worker.post": contract(
+      "delegation.worker.post",
       "post",
       ["worker"],
-      ["sessionId", "runId", "subagentId", "subagentResult"],
-      ["audit.annotate"],
-      ...postBoundary,
-    ),
-    "delegation.background.post": contract(
-      "delegation.background.post",
-      "post",
-      ["worker"],
-      ["sessionId", "runId", "backgroundTaskId", "taskResult"],
+      ["sessionId", "runId", "workerRunId", "workerResult"],
       ["audit.annotate"],
       ...postBoundary,
     ),
@@ -225,18 +209,8 @@ export namespace PolicyPointRegistryModule {
     [Timing.RESOURCES_PREPARE]: ["tool.catalog.pre"],
     [Timing.MODEL_REQUEST]: ["connection.llm.pre"],
     [Timing.MODEL_RESPONSE]: ["connection.llm.post"],
-    [Timing.INVOKE_PREPARE]: [
-      "tool.native.pre",
-      "tool.mcp.pre",
-      "delegation.subagent.pre",
-      "delegation.background.pre",
-    ],
-    [Timing.INVOKE_RESULT]: [
-      "tool.native.post",
-      "tool.mcp.post",
-      "delegation.subagent.post",
-      "delegation.background.post",
-    ],
+    [Timing.INVOKE_PREPARE]: ["tool.native.pre", "tool.mcp.pre", "delegation.worker.pre"],
+    [Timing.INVOKE_RESULT]: ["tool.native.post", "tool.mcp.post", "delegation.worker.post"],
     [Timing.TURN_FINISH]: ["run.turn.post"],
     [Timing.COMPLETION_PREPARE]: ["run.completion.pre"],
     [Timing.WRITEBACK_COMMIT]: ["session.writeback.pre"],
