+++ packages/session/test/worker-run/contract-alias.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test";
-import type { Subagent } from "@openomni/protocol";
+import type { WorkerRun } from "@openomni/protocol";
 import type { WorkerRunStatus } from "../../src/index";
 
 type IsExact<A, B> = (<T>() => T extends A ? 1 : 2) extends <T>() => T extends B ? 1 : 2
@@ -10,11 +10,10 @@ type IsExact<A, B> = (<T>() => T extends A ? 1 : 2) extends <T>() => T extends B
 
 type AssertExact<A, B> = IsExact<A, B> extends true ? true : never;
 
-const workerRunStatusIsProtocolStatus: AssertExact<WorkerRunStatus, Subagent.WorkerRunStatus> =
-  true;
+const workerRunStatusIsProtocolStatus: AssertExact<WorkerRunStatus, WorkerRun.Status> = true;
 
 describe("WorkerRun public contracts", () => {
-  test("WorkerRunStatus intentionally tracks the protocol Subagent.WorkerRunStatus contract", () => {
+  test("WorkerRunStatus intentionally tracks the protocol WorkerRun.Status contract", () => {
     expect(workerRunStatusIsProtocolStatus).toBe(true);
   });
 });
